requirements.txt
llama-index==0.10.28
llama-index-vector-stores-pinecone==0.1.5
llama-index-embeddings-openai==0.1.11
llama-index-llms-openai==0.1.27
openai
fastapi==0.95.2
uvicorn
python-dotenv
